Moved necessary views from new public dashboard to public common
