+++ lib/assets/javascripts/cartodb/new_public_dashboard/entry.js
@@ -4,7 +4,7 @@ var FavMapView = require('./fav_map_view');
 var UserInfoView = require('./user_info_view');
 var PaginationModel = require('../new_common/views/pagination/model');
 var PaginationView = require('../new_common/views/pagination/view');
-var UserSettingsView = require('./user_settings_view');
+var UserSettingsView = require('../public_common/user_settings_view');
 var MapCardPreview = require('../new_common/views/mapcard_preview');
 var LikeView = require('../new_common/views/likes/view');
