+++ lib/assets/javascripts/cartodb/public_common/user_settings/dropdown_view.js
@@ -15,7 +15,7 @@ module.exports = cdb.admin.DropdownMenu.extend({
 
   initialize: function() {
     this.elder('initialize');
-    this.template_base = cdb.templates.getTemplate('new_public_dashboard/user_settings/dropdown_template');
+    this.template_base = cdb.templates.getTemplate('public_common/user_settings/dropdown_template');
 
     // Necessary to hide dialog on click outside popup, for example.
     cdb.god.bind('closeDialogs', this.hide, this);
