+++ lib/assets/javascripts/cartodb/public_common/user_settings_template.jst.ejs
