+++ lib/assets/javascripts/cartodb/public_common/user_settings_view.js
@@ -14,7 +14,7 @@ module.exports = cdb.core.View.extend({
 
   render: function() {
     this.$el.html(
-      cdb.templates.getTemplate('new_public_dashboard/user_settings_template')({
+      cdb.templates.getTemplate('public_common/user_settings_template')({
         avatarUrl: this.model.get('avatar_url')
       })
     );
