+++ lib/assets/test/spec/cartodb/new_public_dashboard/user_settings/dropdown_view.spec.js
@@ -1,4 +1,4 @@
-var SettingsDropdownView = require('../../../../../javascripts/cartodb/new_public_dashboard/user_settings/dropdown_view');
+var SettingsDropdownView = require('../../../../../javascripts/cartodb/public_common/user_settings/dropdown_view');
 var cdbAdmin = require('cdb.admin');
 
 describe('new_public_dashboard/user_settings/dropdown_view', function() {
