+++ lib/assets/test/spec/cartodb/new_public_dashboard/user_settings_view.spec.js
@@ -1,4 +1,4 @@
-var UserSettingsView = require('../../../../javascripts/cartodb/new_public_dashboard/user_settings_view');
+var UserSettingsView = require('../../../../javascripts/cartodb/public_common/user_settings_view');
 var cdbAdmin = require('cdb.admin');
 var $ = require('jquery');
